2025-05-05-aem_uefi.md: update SKL and Xen versions
SKL commit digest changed as the commit is created by CI when the RPM
is built. Xen has been updated on release4.2 branch, which is taken as
base on which TrenchBoot patches are applied. The difference is only
in XSA-469 / CVE-2024-28956, which relates more to runtime code, and
most of AEM happens in setup code.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>